♻️ Extract duplicated test pipeline helpers; fix a silently-wrong hash key
log_skip_summary was defined verbatim in both TestBuildSetup and
TestBuildExecutor, and remove_partials_source_objects verbatim in both
TestBuildPlanner and TestBuildExecutor -- each pair copy-pasted, drifting
independently. Both now live once in TestPipelineHelpers, mixed into all
three classes the same way TestInvokerTypes already is.

Also fixes paths[:preprocess_incudes] to :preprocess_includes -- a typo'd
Hash key silently participates in nothing rather than raising, exactly the
kind of mistake a Hash (as opposed to a Struct) never catches.
